* add params verification

* update error message in params verification

* replace int with uint

* fix dex ut for uint

* update int to uint

* add offset and limit verification

* fix querier page ut

* add x/dex GetPage sanity check

+++ x/order/client/rest/rest_v2.go
@@ -48,7 +48,7 @@ func depthBookHandlerV2(cliCtx context.CLIContext) http.HandlerFunc {
 			return
 		}
 
-		params := keeper.NewQueryDepthBookParams(product, size)
+		params := keeper.NewQueryDepthBookParams(product, uint(size))
 		req, err := cliCtx.Codec.MarshalJSON(params)
 		if err != nil {
 			common.HandleErrorResponseV2(w, http.StatusBadRequest, common.ErrorInvalidParam)
